refactor: redistribute _filesystem_helpers into focused modules
Split the 262-line _filesystem_helpers.py into three focused modules:
- _fs_metadata.py (122 lines): metadata/detection helpers
- _fs_project_helpers.py (86 lines): project-building helpers
- _filesystem_helpers.py (68 lines): dimension helpers only
